% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bayes_estimate}
\alias{bayes_estimate}
\title{Calculate a Bayesian Estimate}
\usage{
bayes_estimate(object)
}
\arguments{
\item{object}{Object of class betabin}
}
\value{
A Bayesian point estimator of theta
}
\description{
Calculate a Bayesian Estimate
}
