% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci}
\alias{ci}
\title{Calculate credible interval for Bayesian estimator}
\usage{
ci(object, ci.level = 0.05)
}
\arguments{
\item{object}{Object of class betabin}

\item{ci.level}{Confidence level}
}
\value{
Lower and upper bound on the credible interval
}
\description{
Calculate credible interval for Bayesian estimator
}
