% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mle_estimate}
\alias{mle_estimate}
\title{Calculate an MLE Estimate}
\usage{
mle_estimate(object)
}
\arguments{
\item{object}{Object of class betabin}
}
\value{
An MLE point estimator of theta
}
\description{
Calculate an MLE Estimate
}
