% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.betabin}
\alias{summary.betabin}
\title{Summarize the model}
\usage{
\method{summary}{betabin}(object, ci.show = TRUE, ci.level = 0.1)
}
\arguments{
\item{object}{Object of class betabin}

\item{ci.show}{A boolean indicating whether credible interval is shown}

\item{ci.level}{Confidence level}
}
\description{
Summarize the model
}
